Background and collision
Set up the background to adjust to the size of the screen and maintain
the color white. Also adjusted the size of the background to support
more people. activated the collision and set it up. It does not work to
send a message to the others that they have been hit. Still trying to
figure out the message setup.